Add t3c calling t3c-verify
